# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simply Plural API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: string;
  name: string;
  pronouns?: string;
  description?: string;
  color: string;
  avatarUrl?: string;
  isFronting: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateMemberInput {
  name: string;
  pronouns?: string;
  description?: string;
  color: string;
  avatarUrl?: string;
}

export interface UpdateMemberInput {
  name?: string;
  pronouns?: string;
  description?: string;
  color?: string;
  avatarUrl?: string;
}

export interface FrontSession {
  id: string;
  memberId: string;
  memberName: string;
  memberColor: string;
  memberAvatarUrl?: string;
  customStatus?: string;
  startTime: string;
  endTime?: string;
  isActive: boolean;
}

export interface StartFrontingInput {
  memberId: string;
  customStatus?: string;
}

export interface UpdateFrontStatusInput {
  customStatus: string;
}

export type GetFrontHistoryParams = {
  startDate?: string;
  endDate?: string;
};