# lib/api-zod/src/generated/types/frontSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simply Plural API
 * OpenAPI spec version: 0.1.0
 */

export interface FrontSession {
  id: string;
  memberId: string;
  memberName: string;
  memberColor: string;
  memberAvatarUrl?: string;
  customStatus?: string;
  startTime: string;
  endTime?: string;
  isActive: boolean;
}