# lib/api-zod/src/generated/types/startFrontingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simply Plural API
 * OpenAPI spec version: 0.1.0
 */

export interface StartFrontingInput {
  memberId: string;
  customStatus?: string;
}