# lib/api-zod/src/generated/types/updateMemberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simply Plural API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMemberInput {
  name?: string;
  pronouns?: string;
  description?: string;
  color?: string;
  avatarUrl?: string;
}